feat: fetch words
